Move a work node's `MutationInfo` to a property of the node, to avoid a bunch of map lookups to find the mutation information for a node.
Also move the logic that resolves the mutations for a node to live in the node, to remove a bunch of knowledge about tasks from the execution plan logic and to allow other kinds of nodes to declare what they will produce.
