+++ subprojects/core/src/main/java/org/gradle/execution/plan/ActionNode.java
@@ -56,6 +56,11 @@ public class ActionNode extends Node implements SelfExecutingNode {
     }
 
     @Override
+    public void resolveMutations() {
+        // Assume has no outputs that can be destroyed or that overlap with another node
+    }
+
+    @Override
     public boolean isPublicNode() {
         return false;
     }
