+++ subprojects/core/src/main/java/org/gradle/execution/plan/LocalTaskNode.java
@@ -16,14 +16,33 @@
 
 package org.gradle.execution.plan;
 
+import com.google.common.collect.ImmutableSet;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
+import org.gradle.api.UncheckedIOException;
 import org.gradle.api.internal.TaskInternal;
+import org.gradle.api.internal.file.FileCollectionFactory;
+import org.gradle.api.internal.file.FileResolver;
+import org.gradle.api.internal.project.ProjectInternal;
 import org.gradle.api.internal.tasks.TaskContainerInternal;
+import org.gradle.api.internal.tasks.TaskPropertyUtils;
+import org.gradle.api.internal.tasks.properties.FileParameterUtils;
+import org.gradle.api.internal.tasks.properties.InputFilePropertyType;
+import org.gradle.api.internal.tasks.properties.OutputFilePropertyType;
+import org.gradle.api.internal.tasks.properties.PropertyValue;
+import org.gradle.api.internal.tasks.properties.PropertyVisitor;
+import org.gradle.api.internal.tasks.properties.PropertyWalker;
+import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.api.tasks.TaskExecutionException;
 import org.gradle.internal.ImmutableActionSet;
+import org.gradle.internal.resources.ResourceDeadlockException;
+import org.gradle.internal.service.ServiceRegistry;
 
 import javax.annotation.Nullable;
+import java.io.File;
+import java.io.IOException;
+import java.util.Map;
 import java.util.Set;
 
 /**
@@ -31,10 +50,12 @@ import java.util.Set;
  */
 public class LocalTaskNode extends TaskNode {
     private final TaskInternal task;
+    private final Map<File, String> canonicalizedFileCache;
     private ImmutableActionSet<Task> postAction = ImmutableActionSet.empty();
 
-    public LocalTaskNode(TaskInternal task) {
+    public LocalTaskNode(TaskInternal task, Map<File, String> canonicalizedFileCache) {
         this.task = task;
+        this.canonicalizedFileCache = canonicalizedFileCache;
     }
 
     @Nullable
@@ -142,5 +163,106 @@ public class LocalTaskNode extends TaskNode {
     @Override
     public String toString() {
         return task.getIdentityPath().toString();
+    }
+
+    @Override
+    public void resolveMutations() {
+        final LocalTaskNode taskNode = this;
+        final TaskInternal task = getTask();
+        final MutationInfo mutations = getMutationInfo();
+        ProjectInternal project = (ProjectInternal) task.getProject();
+        ServiceRegistry serviceRegistry = project.getServices();
+        final FileResolver resolver = serviceRegistry.get(FileResolver.class);
+        final FileCollectionFactory fileCollectionFactory = serviceRegistry.get(FileCollectionFactory.class);
+        PropertyWalker propertyWalker = serviceRegistry.get(PropertyWalker.class);
+        try {
+            TaskPropertyUtils.visitProperties(propertyWalker, task, new PropertyVisitor.Adapter() {
+                @Override
+                public void visitOutputFileProperty(final String propertyName, boolean optional, final PropertyValue value, final OutputFilePropertyType filePropertyType) {
+                    withDeadlockHandling(
+                        taskNode,
+                        "an output",
+                        "output property '" + propertyName + "'",
+                        () -> FileParameterUtils.resolveOutputFilePropertySpecs(
+                            task.toString(),
+                            propertyName,
+                            value,
+                            filePropertyType,
+                            fileCollectionFactory,
+                            outputFilePropertySpec -> mutations.outputPaths.addAll(canonicalizedPaths(canonicalizedFileCache, outputFilePropertySpec.getPropertyFiles()))
+                        )
+                    );
+                    mutations.hasOutputs = true;
+                }
+
+                @Override
+                public void visitLocalStateProperty(final Object value) {
+                    withDeadlockHandling(
+                        taskNode,
+                        "a local state property", "local state properties",
+                        () -> mutations.outputPaths.addAll(canonicalizedPaths(canonicalizedFileCache, resolver.resolveFiles(value))));
+                    mutations.hasLocalState = true;
+                }
+
+                @Override
+                public void visitDestroyableProperty(final Object value) {
+                    withDeadlockHandling(
+                        taskNode,
+                        "a destroyable",
+                        "destroyables",
+                        () -> mutations.destroyablePaths.addAll(canonicalizedPaths(canonicalizedFileCache, resolver.resolveFiles(value))));
+                }
+
+                @Override
+                public void visitInputFileProperty(String propertyName, boolean optional, boolean skipWhenEmpty, boolean incremental, @Nullable Class<? extends FileNormalizer> fileNormalizer, PropertyValue value, InputFilePropertyType filePropertyType) {
+                    mutations.hasFileInputs = true;
+                }
+            });
+        } catch (Exception e) {
+            throw new TaskExecutionException(task, e);
+        }
+
+        mutations.resolved = true;
+
+        if (!mutations.destroyablePaths.isEmpty()) {
+            if (mutations.hasOutputs) {
+                throw new IllegalStateException("Task " + taskNode + " has both outputs and destroyables defined.  A task can define either outputs or destroyables, but not both.");
+            }
+            if (mutations.hasFileInputs) {
+                throw new IllegalStateException("Task " + taskNode + " has both inputs and destroyables defined.  A task can define either inputs or destroyables, but not both.");
+            }
+            if (mutations.hasLocalState) {
+                throw new IllegalStateException("Task " + taskNode + " has both local state and destroyables defined.  A task can define either local state or destroyables, but not both.");
+            }
+        }
+    }
+
+    private static ImmutableSet<String> canonicalizedPaths(final Map<File, String> cache, Iterable<File> files) {
+        ImmutableSet.Builder<String> builder = ImmutableSet.builder();
+        for (File file : files) {
+            builder.add(canonicalizePath(file, cache));
+        }
+        return builder.build();
+    }
+
+    private static String canonicalizePath(File file, Map<File, String> cache) {
+        try {
+            String path = cache.get(file);
+            if (path == null) {
+                path = file.getCanonicalPath();
+                cache.put(file, path);
+            }
+            return path;
+        } catch (IOException e) {
+            throw new UncheckedIOException(e);
+        }
+    }
+
+    private void withDeadlockHandling(TaskNode task, String singular, String description, Runnable runnable) {
+        try {
+            runnable.run();
+        } catch (ResourceDeadlockException e) {
+            throw new IllegalStateException(String.format("A deadlock was detected while resolving the %s for task '%s'. This can be caused, for instance, by %s property causing dependency resolution.", description, task, singular), e);
+        }
     }
 }
