+++ subprojects/core/src/main/java/org/gradle/execution/plan/MutationInfo.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.execution.plan;
+
+import com.google.common.collect.Sets;
+
+import java.util.Set;
+
+class MutationInfo {
+    final Node node;
+    final Set<Node> consumingNodes = Sets.newHashSet();
+    final Set<Node> producingNodes = Sets.newHashSet();
+    final Set<String> outputPaths = Sets.newHashSet();
+    final Set<String> destroyablePaths = Sets.newHashSet();
+    boolean hasFileInputs;
+    boolean hasOutputs;
+    boolean hasLocalState;
+    boolean resolved;
+
+    MutationInfo(Node node) {
+        this.node = node;
+    }
+}
