+++ subprojects/core/src/main/java/org/gradle/execution/plan/TaskNodeFactory.java
@@ -17,6 +17,7 @@
 package org.gradle.execution.plan;
 
 
+import com.google.common.collect.Maps;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
@@ -30,6 +31,7 @@ import org.gradle.internal.Actions;
 import org.gradle.internal.build.BuildState;
 
 import javax.annotation.Nullable;
+import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
@@ -39,6 +41,7 @@ public class TaskNodeFactory {
     private final IncludedBuildTaskGraph taskGraph;
     private final GradleInternal thisBuild;
     private final BuildIdentifier currentBuildId;
+    private final Map<File, String> canonicalizedFileCache = Maps.newIdentityHashMap();
 
     public TaskNodeFactory(GradleInternal thisBuild, IncludedBuildTaskGraph taskGraph) {
         this.thisBuild = thisBuild;
@@ -54,7 +57,7 @@ public class TaskNodeFactory {
         TaskNode node = nodes.get(task);
         if (node == null) {
             if (task.getProject().getGradle() == thisBuild) {
-                node = new LocalTaskNode((TaskInternal) task);
+                node = new LocalTaskNode((TaskInternal) task, canonicalizedFileCache);
             } else {
                 node = new TaskInAnotherBuild((TaskInternal) task, currentBuildId, taskGraph);
             }
@@ -171,6 +174,11 @@ public class TaskNodeFactory {
             }
             TaskInAnotherBuild taskNode = (TaskInAnotherBuild) other;
             return task.getIdentityPath().compareTo(taskNode.task.getIdentityPath());
+        }
+
+        @Override
+        public void resolveMutations() {
+            // Assume for now that no task in the consuming build will destroy the outputs of this task or overlaps with this task
         }
 
         @Override
