+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/TransformationNode.java
@@ -80,6 +80,11 @@ public abstract class TransformationNode extends Node implements SelfExecutingNo
     }
 
     @Override
+    public void resolveMutations() {
+        // Assume for now that no other node is going to destroy the transform outputs, or overlap with them
+    }
+
+    @Override
     public String toString() {
         return transformationStep.getDisplayName();
     }
